Get list of header starts and footer starts, Save carved bytes to file

+++ app/src/main/java/com/daksh/scalpelandroid/storage/room/entity/Rule.kt
@@ -11,9 +11,9 @@ data class Rule(
 
         val extension: String,
 
-        val minBytesAmount: Long = 0,
+        val minBytesAmount: Int = 0,
 
-        val maxBytesAmount: Long,
+        val maxBytesAmount: Int,
 
         val header: String,
